A whole lot of new stuff

+++ src/unistd/write.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <unistd.h>
 
-ssize_t write(int fd, const void *buf, int len)
+ssize_t write(int fd, const void *buf, size_t len)
 {
     return (ssize_t)syscall(SYS_write, fd, buf, len);
 }
